팀장 - ㅋㅁ

+++ MDS/src/main/java/kr/mds/entity/tb_Security_alarm_car.java
@@ -1,5 +1,6 @@
 package kr.mds.entity;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 import lombok.AllArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class tb_Security_alarm_car {
-	private int sac_num;
+	private Long sac_num;
 	private String cctv_id;
 	private String car_num;
 	private String car_inout_at;
